Add day 2
This was quite fun. Using internal representations of types to represent the
scores I found really enjoyable, though that got in the way a bit when I needed
to calculate the winning/drawing/losing move, as the nicest way of doing that
is to use modular arithmetic!

I figured at that point it was nicer to just add a score method instead.