Add day 4
I was really chuffed with today's solution - I went ahead and implemented a
general overlap function first, and that went a long way to figuring it out. The
fact that sections are a value type so that equality on the values themselves is
super easy was really nice, too.

Using the generic Pair type rather than just making Assignments first class was
completely unnecessary, but it's nice seeing a little bit of how Go generics
work.