Add part 2
My original solution didn't handle edges properly... it was just two -1 factors
missing

I think the problem was that otherwise it'd attempt to go past the grid. Of
course, the edges of the grid don't actually matter how tall they are - either
the view was already blocked, or the final square is the last one seen.