Add day 12
Probably implementing a PriorityQueue from scratch was... not necessary, but
it's been a very long time since I last did so, and it ain't half a good way of
using Go's pointer functionality.

It almost felt like writing C, even - one of my bits of uni coursework was
something very similar - yet the language is much nicer to use.