[day16] Optimise part2 by only following shortest paths
This now completes the toy example very quickly (and correctly!), but still can't solve the real part2

Reducing the number of states (by collapsing the useless flowrate 0 tunnels) seems to be the next step, but it doesn't work well with how I'm memoising previously-encountered states.

On the other hand, it avoids the situation I feel like I'm in at the moment, where we consider situations where the actors just move backwards and forwards along tunnels, wasting time (i.e. we're not actually committing to the path being taken)