Implement day16 again
The original day16 folder will remain as testament to the pain that I had with this challenge.

Honestly though, I'm glad I persevered - even if it did take a bit of external observation to get to a solution. I feel like I understand dynamic programming in general a bit better now, even if I'm unlikely to come across a graph problem like this in the real world. And I especially understand writing recursive Go code better!

I'm sure this solution still has room for optimisation, but the fact it completes in a reasonable amount of time is good enough for me.