[day19] Tighten best-case bound to t(t-1)/2
There was an off-by-one error here: yes, we could build a geode bot each turn,
but it won't start producing until the end of the subsequent turn. As such, we
can actually prune slightly sooner

With a time go run ., this runs about 10 seconds quicker on my machine, going
from 35s to 25s. Not a necessary improvement by any means, but I wanted to
correct the error.